AAM-34808: Add polyfills.js that includes regeneratorRuntime
This specific polyfill is needed to use the `TableViewDataSource`
and `Autocomplete` react-spectrum components, and possibly otheres.

See this slack thread for more details:
https://adobetechops.slack.com/archives/C7N5BE1EW/p1509986884000764

Apps made with create-react-app contain polyfills (and I think
babel-polyfill), but they just need to be imported as needed.
That would explain why importing `regeneratorRuntime` just works,
without having to add any extra polyfills or packages to the project.
See this SO thread for more info:
https://stackoverflow.com/questions/43756211/best-way-to-polyfill-es6-features-in-react-app-that-uses-create-react-app