Optimize redux-devtools-extensions for production use
Using `logOnlyInProduction` allows us to have full functionality
of the extension in a development environment, and restricts to
basic logging functionality in production. This means that we do
ship the extension in prod, but then it allows us to debug state in
production with a minimal footprint, and the extension is paused
unless the window is open, so it shouldn't slow users down.

See more info here:
https://medium.com/@zalmoxis/using-redux-devtools-in-production-4c5b56c5600f