AAM-34808: Update failing snapshot related to TableViewDataSource
TL;DR please update to the latest version of node (not LTS).

The react-spectrum `TableViewDataSource` class extends the Adobe
`React/collection-view`'s `DataSource` class, which extends
node's native `EventEmitter` class inside the `events` module.
That class had a property called `domain` on it that got
deprecated somewhere in between node versions 8.4.0, which I was
running locally, and 9.6.1, which Jenkins was running. Therefore
the `domain` property was showing up in my snapshots that used
components that extend `TableViewDataSource`, but they weren't in
the Jenkins compiled app so there was a mismatch with the snapshot
I pushed.

After upgrading my local node to the latest (9.7.1), my DataSource
instances no longer had the `domain` property and I was able to
update the snapshot. When I push this to Jenkins, it will run 9.6.1
which also expect no `domain` property to exist and the snapshot
should match.