AAM-34808: Modify table content & styles to fill 100% width
There's a small bug where the table initially flashes/renders
without the Adobe fonts loaded--during this time the column headers
are correctly lined up with their cells. But once the Adobe font loads
and the table re-renders with it, the widths get messed up and the
cells are roughly 20px to the left of the headers. Clicking any
of the column headers fixes this.