AAM-36054: Generate keys in en.json per component
Use the library `extract-react-intl-messages`, which uses
`babel-plugin-react-intl` to parse all the js files in the
project and put together en.json with all of the messages.

Key ids are named like the filepath to the file,
delimited by periods and ending with the name of the
message (also the name of the actual object key). See code
for example. Message data is defined using `defineMessages`
and then spread into the appropriate `<FM>` tag.

Removing country codes from now since this library doesn't
currently use them, and I don't think we will need them.