AAM-36047: Add colored bar visualization for percentage change
Next to the actual percentage in each cell, show a colored bar
that extends from a middle divider. A green bar extends to the
right for a positive % change, and a red bar extends to the left
or a negative % change.

The width of the colored bar is normalized so that in any given
rendering of the SignalsTable, the percentage change with the largest
magnitude has the largest possible bar size of 2rem. There is also
a min width enforced so super small percentages still show up.