AAM-36047: Add test helper for Intl components and test percentage number
In order to test the fully rendered output of a component that
uses a react-intl component (like FormattedNumber), we need to add
some helper methods that wrap these react-intl components with
an appropriate IntlProvider, so it can pass the intl data as a
context. I adapted platform UI's solution, which closely resembles
the react-intl docs' suggestion. See platform's helper file here:e

https://git.corp.adobe.com/experience-platform/platform-ui/blob/blackout/test/helpers/componentHelpers.js