Revert "AAM-34809: Use css module @value export for maxBarWidth variable"
This reverts commit a4447511b99ea0cb0d7871b5fd513d70c67812ff.

This resulted in a few failing tests because jest tests don't correctly
import css-modules by default. I tried solving it by updating
to react-scripts version 2.0.0-next, but ran into problems and didn't
want to spend much time on it. In this release, CRA provides opt-in support
for CSS modules. We should use it as soon as it's available.