AAM-36340: Remove encodeKeyValuePairs and add stringifySignals
We are trying a new approach; rather than encoding the key value
pairs inside the URL as query string parameters, we will (1)
stringify them in the format needed for trait rules and (2) store
them in sessionStorage.

The method `encodeKeyValuePairs` was doing two things: stringifying
the key value pairs and then preparing them for storage in a URL.
The new `stringifySignals` method just takes care of the first part.
I will write separate code for storing and retrieving data.