AAM-36340: Add utils/urls.js for navigating to other AAM urls
For now, this only holds URLs for the Trait Create page (rule-based
and onboarded versions). Also adds a `serialize` method for
query string params, that might be safer to switch to the
`query-strings` npm module.

In order to correctly target the Portal base URL (aka the old app),
I added environment variable files that define the portal's base
URL. In development, it's on localhost:8080--since Signals is on
localhost:3000, we can't use relative hyperlinks. But in production
they'll be reached from the same server, so we can, and the env
variable accounts for that. Setting up env variables in these named
files and with the necessary "REACT_APP_" prefix comes out of the
box with CRA.