AAM-36340: Generate rule-based/onboarded create trait url
Did some refactoring to allow determining whether to generate the URL
for the Create Rule-Based Trait or Create Onboarded Trait page.

Removed the `traitsCreationLabelText` prop from the `TraitsCreation`
component and moved it down one level so it's only passed into the
`SingleSignalTraitsCreation` component. The `MultiSignalsTraitsCreation`
component will always hard code its button label so we don't need to
pass it in as a prop to their parent component, `TraitsCreation`.