AAM-36340: Move 'dataType' logic for multi signals to redux selector
Making this a selector rather than putting it in the component code
or altering the reducer shape allows us to consider the `dataType`
as a computed property of the selected signals in reducer state,
meaning it's able to be calculated from existing properties and is mainly
helpful for the UI. We wouldn't want to modify the actual shape
because then it would get out of sync with the API structure,
which the reducer is meant to reflect.

I wasn't able to do the same with the added `dataType` property of the
actual `results` reducer shape as I just described above, because
calling `.map()` on an array prop returned from such a selector returns
a new reference and fires another render on `SearchContainer`, which
eventually re-renders the `SignalsTable`--we run into that error
where selecting a row immediately re-renders the table so the items
get overwritten and the checkbox disappears.

the next commit I will do the same thing for the `dataType`
property on the results reducer list in addition to the selectedSignals.