AAM-36340: Remove `createTraitFromMultiSignals` action and `handleTraitsCreation` handlers/props
Since we went the URL and not the sessionStorage route, all the prep
for navigation happens in `TraitCreation#getCreateTraitURL()`, and
we simply pass the correct URL as the `href` of the anchor tags
surrounding the Create Trait link/button. We don't need an action,
and this one was here as a thoughtful placeholder.