AAM-37189: Add custom start and end date datepickers
These will only appear when the user selects "Custom Date Range"
in the "View Records For" dropdown. These date ranges will be
stored in new `customStartDate` and `customEndDate` fields in
the `SearchContainer` state. Eventually, we'll modify
`normalizeSearch` so that if "custom" is selected, we'll parse
the `customStartDate` and `customEndDate` as timestamps and
send those in `startTime` and `endTime` to the API.