AAM-37189: Implement getDaysAgoTimestamp for use in normalizeSearch
We will call this on either the value in the View Records For
dropdown or the custom start/end date datepickers on the search
page in order to normalize a date range preset (ex: '7D') or
a date (ex: '2018-04-24') into a timestamp. The timestamp is
calculated on the date as described above, and at the exact
time of day as the execution of the method by using moment.utc().