AAM-37189: Normalize start/end dates in API requests
When "View Records For" is a date range preset (ex: 'Last 7 Days'),
`startDate` is a timestamp at that many days ago at the time of
day that the method is invoked (ex: '7 days ago, right now'). The
`endDate` is null.

When "View Records For" is "Custom Date Range", `startDate` is a
timestamp at the chosen custom date at the time of day that the
method is invoked (ex: 'April 15, 2018, right now'). The `endDate`
is a similarly calculated timestamp at the chosen custom end date.

Refactored other normalizing methods in `normalizeSearch` and added
test cases for all normalizations.