AAM-37189: Hook up "View Records For" to saved searches
When saving a search, save the `viewRecordsFor`, `customStartDate`,
and `customEndDate` properties from the `SearchContainer` state.
We don't need the `startDate` and `endDate` properties, which are
calculated in `normalizeSearch` right before a search API call.
We also want to save a date range preset like "Last 7 Days" rather
than the exact start and end times that corresponded to "Last 7
Days" at the time it was searched and saved.

When viewing a Saved Search Popover, display either "Last X Days"
or the custom date range by the label "View Records For". Use
`react-intl` for formatting.

When executing a saved search, populate all the search forms
correctly with the `viewRecordsFor` and custom start/end date
properties, which then get normalized into `startDate` and
`endDate` in `normalizeSearch` as usual.