AAM-37189: [fixup] Fix snapshot date mocking
I created the `enzyme` wrapper for the `CustomDateRange` component
inside of the `describe` instead of the `it` block, so setting
`mockdate` in `beforeEach` wasn't taking effect and the wrapper
that was used in the snapshot test was using today's date. That
unit test passed yesterday but failed today.

Moved the wrapper instantiation to the 2 `it` blocks inside that
`describe` block.