AAM-37189: Handle user inputs in datepicker component
The datepicker change handlers inside `SearchContainer` were
assuming a moment object argument, since that's returned by
the <Datepicker /> components' handler when the user selects
them in the calendar. But this broke when the user typed in
dates manually into the datepicker.

Now, we parse just the `valueText` argument of the change
handler, which is either the date selected in the calendar
in the format 'MM/DD/YYYY', or whatever the user types into
the datepicker.