Fix failing Cypress test related to View Records For dropdown
TL;DR
This fix updates the fixture json with the correct `viewRecordsFor`
field of "7D" to match the now-correctly rendering expected prop on
the View Records For dropdown.

Long version
This test looks for the View Records For dropdown, which we give
an prop of `[data-test="view-records"]`. Because of a bug in
react-spectrum that was only fixed recently, this prop wasn't
actually passed down to the `<Button />` rendered by the
`<Select />` component, so the spec read the prop as `undefined`.
Coincidentally, the fixture data for saved searches was out of date
and didn't include the updated schema that we save to the User
Annotation API: we send `viewRecordsFor`, `customStartDate`, and
`customEndDate` rather than `startDate` and `endDate`. So when the
spec was looking for the `viewRecordsFor` field in the fixture data,
it was also `undefined`. Both sides of the equation were returning
`undefined` so the test passed!

Luckily, react-spectrum just fixed the issue on their side, so the
`<Button />` component rendered by the `<Select />` now receives
`{ ...otherProps }`, which now includes our `data-test`--in the case
of our tests, it's always set to "7D". Now we just update the
fixture json to correctly be "7D" so the test passes.