AAM-37564: Remove proxy and homepage
From this point on, the app is intended to run on localhost:8000
or in an env that has the lambda architecture set up behind the
same URL, so that calls to `/portal/api/v1/...` hit the portal API.

It would be nice in the future to still be able to run the dev
server on :3000 and make calls to json-server. But for now we
rip the band aid off!