AAM-37811: Update report suite friendly name autocomplete endpoint
I was using the endpoint `/report-suites/dimensions/${rsId}`, which
was a WIP endpoint. The real one is now `/signals/keys`.

In addition to passing the report suite ID as a query param to this
endpoint, I also modified `getKeysByReportSuiteId` to accept a
`key` param, which is the user input into the field, so that the
field still acts as a search/autocomplete.