AAM-37793: "View Records For" dates normalize to UTC midnight
When selecting an option from the "View Records For" dropdown
(i.e. either "Last X Days" or a custom date range), the time of day
we will send to the API for the start (and optional end) date is
midight UTC of the selected day. We were previously sending the
current time of day (i.e. "now, 7 days ago"), but after discussion
with product, users are more used to the idea of "data from the
last '7 days plus'"--7 full business days, plus today's UTC midnight
until right now.n