AAM-37793: Add integration tests for normalizing start/end dates
Verify that when selecting each "Last X Days" option, `startDate`
is normalized to X days ago at UTC midnight in ms and endDate is
normalized to `null`.

Verify that when selecting a custom start and end date, `startDate`
and `endDate` are normalized to the selected dates at UTC midnight
in ms.