Fix inconsistencies with `sourceType` when "All" filter selected
In UI code, when the "All" signal type filter is selected, we
will track it with the value `"ALL"` in all places. Anywhere
that we need to deal with values normalized for the API (aka
in `normalizeSearch`), we convert it to `null`, not the empty
string.

Bug explanation:
We were getting lucky before because <SearchContainer > had
initial state where `sourceType` was `null`. Executing a search
from the UI would work. Then if you switched the signal type
filter to anything else and then back to "All", it would then
translate `"ALL"` into `''`, resulting in a 400.