Migrate to react-spectrum Provider for styles
Removed the deprecated `import '@react/react-spectrum/page'` and
instead import `Provider` as a tag to wrap the app. It basically
makes a root element that contains "spectrum spectrum--light" styles,
so that the `body` tag isn't polluted with them.

I added a style of `overflow: 'auto'` to the top-level `<Grid />`.
This fixes styling issues of the top and bottom of the page getting
a white background underneath the margins of the top-most and
bottom-most elements.