AAM-38243: Enforce medium scale in prod builds
TL;DR: Everything looks really big in production builds because
we need to specify the "scale" of the app in a config.  Added a
hacky library that allows us to modify webpack configs without
ejecting CRA, so that we can specify "medium" scale:
http://react-spectrum.corp.adobe.com/beta/guides/getting_started

Long version:

React-spectrum 2.7.0 shipped with a change that requires users to
specify desired values for themes (light/dark etc) and scales
(medium or large) in the form of env vars. The main consequence
of this change, which is biting us now, is that if none of these
configs are specified, then ALL are loaded. In production builds,
every component gets a `index.css` and `index-lg.css` file loaded.
Since `index-lg.css` is loaded second, its styles override
`index.css` for every single component, effectively giving the app
the "large" scale.

The way react-spectrum recommends specifyign the default configs
to avoid this is in the webpack config, within a new `DefinePlugin`,
as seen here:

http://react-spectrum.corp.adobe.com/beta/guides/getting_started

However, users of create-react-app (like us) don't have access to
this webpack config without ejecting, which we'd like to avoid as
long as possible. Furthermore, all custom env vars in CRA apps are
prepended with `REACT_APP_` (ex: `REACT_APP_SCALE_MEDIUM`), which
react-spectrum doesn't support.

After spending a good while trying to add CRA support in react-
spectrum, I found that it's not really possible without a larger
rethinking of how they dynamically import CSS files. That left us
with two options: eject and modify our webpack config, or use the
kinda-community-approved library for modifying react-scripts configs
without ejecting, `react-app-rewired`. It allows us to just add the
`DefinePlugin` bit we need and everything works!

This solution is definitely not ideal, but it's more maintainable
than maintaining our own hacky fork of CRA or ejecting, and if
react-spectrum ever does fix the issue on their side, we can easily
revert this change without trying to "undo" an eject.