AAM-38243: Remove Jest CLI, effectively downgrading jest to 20.0.4
I ran into this issue while trying to solve AAM-38243.

Jest CLI is actually not supposed to be installed in CRA apps,
because it installs a conflicting (higher) verison of Jest. We were
getting an upgraded version of Jest for free by having jest-cli
installed, but trying to integrate it with react-app-rewired for
AAM-38243 broke all our unit tests.

This PR does what the FB team recommends and only using CRA's version
of Jest. By downgrading, we lose some functionalities included in
2 major releases, including `jest.restoreAllMocks`() (accounted for
in this PR) and pretty formatting of snapshots (accounted for in
the next commit). But this is a good trade off for being able to
incorporate react-app-rewired and fix AAM-38243. Once CRA updates
to 2.0.0, we'll have a newer version Jest back. Sorry!