AAM-38243: Update snapshots to format ouput by jest 20.0.4
These snapshots are uglier and have worse indentation formatting.
But we need to deal with them for now until CRA updates to a
later version of jest.

As mentioned in the previous commit, the reason we had more nicely
formatted snapshots is because they were being generated by a later
version of jest included in jest-cli. In the previous commit, I
removed jest-cli, defaulting us to CRA's older version.