AAM-37267: Add actions/limits.js and traitBackfill.reducer.js
Start to add logic to handle the `maxSignalRetentionDays` limit.
Added a new action to fetch limits and a new reducer to store just
the retention days limit. The plan is to refactor the action/reducer
for `getSavedSearchLimit()` so that the saved search limit reducer
handles the more generic `FETCH_LIMITS_FULFILLED` action but keeps
the same reducer shape (keep track of the saved search limit in
the `savedSearch` reducer).