AAM-38249: Debounce `loadMore` calls
Inside `<SearchContainer />`'s `handleLoadMore` function, we add
an action call that updates an `isThrottled` reducer property.
Then we return early if we see that `isThrottled` is true.
After we call `loadMore()` to fetch the next page of results,
we debounce the action call that disables the throttle. That way,
if a heavy scroll event triggers lots of `handleLoadMore` calls,
all of them except the first one will return early before actually
calling `loadMore`.