AAM-38453: Show error message if all traits unavailable
This is most likely to happen if the traits are forbidden (403),
not so much unavailable (404) or because of a bad request (400),
so we optimistically mention permissions in this error message.

And as another step towards only surfacing errors when absolutely
necessary, if the # of successful trait fetches is fewer than
`maxVisibleTraits` and some trait calls errored, we lump those
errored ones in the "And X more" message that also shows when
the number of successful trait fetches is greater than
`maxVisibleTraits`. We also do this rather than always showing an
error message when some calls fail, because to accurately display
how many traits are unavailable/forbidden, we'd need to make GET
calls for every single one, which is inefficient. d