AAM-38428: Fix checkbox selection in Signals Table
Implement `shouldComponentUpdate` in the common `Table` component
that checks if the selected signals have updated--if they have
(meaning that the only state change was triggered by the user
selecting a checkbox), we should not re-render the table. The
table re-rendering was what was making the checkboxes appear to
not be selected.

This was accomplished by including the `selectedRowIndex` array
in the entire redux loop: it's updated in redux state when the user
selects a checkbox, and then we pass that state all the way down
through `SignalsTable` and then to `Table` to perform this check.