AAM-37861: Lazy load no longer jumps to top of table
Refactored the `<Table />` component to create its own `DataSource`
in the `constructor` and then imperatively update the items and
columns etc in `componentDidUpdate`, instead of creating a brand new
`DataSource` inside `render`, which basically wiped the Table's
state clean on each re-render.

`componentDidUpdate` is the recommended lifecycle method for calling
3rd party libraries that need to update the DOM separately from
React's virtual DOM, which is our exact use case here.