AAM-38132: Handle LIST saved search 404 silently
If the saved search LIST request returns a 404, that means the user
hasn't created any saved searches yet, and the annotation in the DB
doesn't yet exist for that user. Since this is the expected behavior
for users who access Signals for the first time, we shouldn't show
an error, and should instead just rely on the fact that users will
always have the saved search presets, and will never have an empty
list of saved searches.