AAM-38132: Restore `status` to createAsyncAction error payload
I basically undid all my work in #134 with the commit d5d4466. TIL
that spreading a fetch response, which is a `Response` object and
not a JSON object, returns an empty object. By returning `...response`
inside the payload instead of just the original `response` object,
I lost the `status` field, which meant the error reducer couldn't
properly check if the error was a 404 or not, and it showed the
old error message again.