AAM-38971: Store signals results sort opts in redux, use in loadMore
In order to access the current sort options inside `loadMore`,
a function defined in `SearchContainer`, we need `SearchContainer`
to have access to the currently sorted column and direction. To do
this, create new state in the `results` reducer to track the
`sortBy` and `sortDir` properties generated by clicking table column
headers.