AAM-39094: Allow custom start/end dates to be same date
Adjust the boundaries of the custom date pickers' min and max props.

Now that the custom end date is normalized to the EOD of the selected
date, users should be able to select the same date for the start
and end dates, since this will now represent the range of one full
day on the selected date.