Fix redux-actions/node_modules/camelcase breaking build
Earlier today, the redux-actions library just released an update
that removes all of its lodash dependencies--one of them was
lodash/camelcase, which they've substituted for the camelcase npm
module. lodash/camelcase is written in ES5, but the camelcase module
is written in ES6/7. create-react-app expects imported node modules to
be written in ES5—its uglifier breaks when trying to uglify ES6/7 code.

Fixed this by locking the version to the latest release, v2.5.1, which
imports a version of lodash/camelcase written in ES5.