AAM-38970: Include sort options when executing saved search (#159)
* AAM-38970: Fix `getSortOptions` selector

The sort options are stored in redux state as the normalized
properties `sortBy` and `descending`, but the selector was returning
`sortBy` and `sortDir`--`sortDir` is only used by the table.

* AAM-38970: Move `normalizeSortOptions` from `normalizeSearch` to `handleSortSearch`

The handler `handleSortSearch` is the only place where we actually
need to normalize the sort options provided by the table before
doing things with them. In `handleLoadMore` we're accessing the
normalized options from redux state. And in `handleSavedSearchClick`
we're accessing normalized options as saved in the annotations data.

* AAM-38970: Correct saved search preset sort options

Both saved search presets use the API's default sort options:
descending and by `totalCount`. Both of these were wrong in the
preset constants file, and since the API uses the right ones by
default, we can just omit them altogether.

* AAM-38970: Fix cy test and show "Total Counts" in saved search popover

Forgot to update an integration test that checks the selected sort
column in the saved search popover. Decided to explicitly put
`sortBy: totalCount` in the presets file to make sure it appears
in the popover.

* AAM-38970: Fix another failing cypress saved search test

* AAM-38970: Update snapshots