Topnav: Render topnav (without icons) in local environment
Instead of copying the topnav code to the build/static folder after
`npm run build`, we copy it to `public/static` after `npm install`.
That way we have access to it in a development environment and not
just with production builds hosted at `npm run start:ci`.

This works because of a weird combination of:
- Apache rewrite rules for /portal/signals
- Lambda authentication rerouting `/static` to `http://localhost:3000/static`
  - We still get a CORS issue for icons locally due to this
- Create React App's build step copying everything from `/public` to `/build`

It's hacky but it works for now.

Note for @parga:
Also added --unsafe-perm to Dockerfile's `npm install` so that
the "postinstall" script fires. Otherwise it won't fire when ran
by the root user. Not sure if this is the best approach or if we
should create a non-root user in the Dockerfile.