Max row height is 72px, calculated from results instead of prop
React-spectrum defines a max rowheight of 72 in the `TableView`
component. Modify our logic for calculating row and table heights
so that if signal results have 1 KVP the rows are the default height
of 48px and if they have 2 or 2 KVPs the rows are the max height of
72px. It looks a bit squished with 3 KVPs but it's what react-
spectrum enforces.

The `SignalsTable` component also no longer needs to receive the
`totalKeyValuePairs` prop and calculates this property from the
lengt of `keyValuePairs` in each signal result. This also fixes
a bug where clicking "Add"/"Remove" in the search form would
expand/contract the rows in the table without actually performing
any new search.