Dashboard table height exactly fits number of visible rows
There was a small discrepancy between the signals tables on the
dashboard and search pages. The search page table has a header
row height of 40px because it contains a checkbox, but the dashboard
table height has a header row height of 20px because it has no
checkbox. Account for this in `getTableHeight()`.