AAM-39715: Show loading spinner in between searches (#187)
* AAM-39715: Update `createAsyncAction` to better handle PENDING state

TL;DR: Dispatch `_PENDING` actions before, and not after, API requests.
No longer use `redux-promise-middleware` to handle API-calling actions.

LONG VERSION:

The logic in `createAsyncAction` was doing the following:
- Wait until initial `fetch` call returns (which takes basically
  the entirety of the async call)
- Dispatch action with the given type (ex: `FETCH_THING`)
- Rely on `redux-promise-middleware` to then dispatch
  `FETCH_THING_PENDING` and `FETCH_THING_FULFILLED/REJECTED`,
  depending on if `response.ok` is true --> calling response.json()

That means that `FETCH_THING_PENDING` is called after the API call
returns (either successfully with 2xx/3xx or with error with 4xx/5xx).
We don't actually dispatch `FETCH_THING_PENDING` while the API
request is in flight, but actually only in between getting the
initial response and trying to call `response.json()`. That's partially
why our loading spinners weren't working so well.

The solution is to bypass `redux-promise-middleware` altogether
and use `createAsyncAction` as a full custom solution that's tailored
to work with `fetch`. When the action `FETCH_THING` is fired, don't
dispatch it (like before) and instead immediately dispatch
`FETCH_THING_PENDING` (this is new). Then, await the API call's
response. If `response.ok` is true, await `response.json()` and dispatch
the result of that as the payload of `FETCH_THING_FULFILLED`. If
`response.ok` was false, dispatch `FETCH_THING_REJECTED` with our
custom error schema (like before). If for whatever reason calling
`response.json()` or anything else had failed (which shouldn't
really happen), dispatch `FETCH_THING_REJECTED` with the error as
the payload (like before).

Later it'll be worth it to remove `redux-promise-middleware` after
we inspect if we really don't need it anymore. We really need something
like `redux-fetch-middleware`, but all of the available libraries
look too over-engineered and coupled to custom redux action schemas
that I'd rather not tie us to. Maybe as a small step we can port
`createAsyncAction` to a custom middleware for us to use in Signals.

* AAM-39715: Show loading spinner in between searches

The loading spinner now appears in between searches, and not just
while the first search is loading. This makes it look less like
the table/page is frozen after firing subsequent searches.

* AAM-39715: Simplify logic for checking if results exist

Remove Boolean check and Object.keys() call on this array. Do the
same in the render logic that shows `<SaveSearchExecution />`.

* AAM-39715: [bugfix] Fix state modification of shared wrapper

This unit test was changing the `errors` props of the global `wrapper`
in a way that deleted all properties besides the `searchForm` errors.
This change propagated to all later tests. Spread the existing
`errors` props so that future tests have the other error properties.

Making this fix so I can add another unit test without it breaking,
since it's looking for `hasError` of undefined.

* AAM-39715: Add unit tests for `shouldShowResults` and `isSearchDisabled`

* AAM-39715: Added integration tests