AAM-39857: Empty selected rows in redux after calling search or lazy load (#190)
* AAM-39857: Fix this flow: Select row > search > select same row

Every time the user performs a search (especially after selecting
some rows), we now reset the `selectedRowIndexes` state to an empty
array to reflect that the table component has reset all rows itself.

This fixes an edge case bug as described in the commit name. What
was happening was the `selectedRowIndexes` stayed the same from
pre-search to post-search, so `shouldComponentUpdate` for the
`<Table />` component was returning true. The checkbox would be
unclickable the seocnd time because it's turning on and quickly off.

* AAM-39857: Clear signal selections after lazy load

Previously, selecting a signal and then loading more would show
that the signal is still selected above the table. But as soon
as you selected another signal, that message would reset to 1
signal selected again because the table actually cleared the
selection while the state had stayed at 1 selection and then went
again to 1 selection. This basically implements "we don't support
selecting signals and then lazy loading" which is a problem the
react-spectrum team has faced too.

* AAM-39857: Add comment explaining shouldComponentUpdate
p