AAM-39767: Restore lazy loading on dashboard page (#193)
* AAM-39767: Restore dashboard lazy loading with topnav

These changes are thanks to @elwang. I extracted `isBottomPassed`
to a util and mocked it in the dashboard jest file (it was always
returning true before) to avoid a huge snapshot update.

* AAM-39767: Restore automatic lazy loading if bottom is visible

This fixes some failing integraiton tests.

Now that we've added the topnav and footer, we needed to adjust
the `min-height` of the top-level `<Grid />` element to fit neatly
in between the two. This way, if the bottom of the `<Grid />` is
visible (which, after this fix, now happens again by default if
the screen is zoomed out a lot), we lazy load more tables.