AAM-40002: Throttle dashboard lazy loading to 1 saved search at a time (#195)
* AAM-40002: Throttle dashboard lazy load scroll handler

Scroll events are best optimized by throttling them. This ensures
that it feels snappy by executing as soon as the first scroll is
triggered, and also gets executed every so often at a reasonable
rate, unlike debouncing which would wait until some ms after the
very last scroll event to fire, which feels laggy.

Updated `componentWillReceiveProps` to `componentDidMount`.

* AAM-40002: Rename `isBottomPassed` --> `isNearBottom`

It felt smoother to trigger the lazy load function when the user is
100px from the bottom than when they're actually at the bottom.

* AAM-40002: Refactor lazy load logic

Moved the logic in the `lazyLoad` util into DashboardContainer and
deleted the util. Simplifed the logic by moving the saved search
lists' lengths comparison to a selector.

* AAM-40002: Update int tests to expect 1 lazy loaded table at a time