AAM-41407: Refactor getting shared dataSourceIds of all selected Onboarded signals (#217)
* AAM-41407: Refactor getting shared dsIds of all selected OB signals

Added a few selectors, notably `doAllSelectedResultsShareSameDataSource`
and `getSelectedResultsDataSources`.

For single signal creation, keep the approach of checking the
signal result's `signalType` prop--if it's an onboarded signal,
then store the `source` in sessionStorage.

For muti signal creation, we no longer rely on the `signalType`
prop to determine whether we should store the `dataSourceIds`
of the selected signals. That's because it's tied to the signal
type filters, and users could still achieve this use case with the
"ALL" signal, not just the "ONBOARDED" signal.

Secondly, fixed the business logic so that we only store the `source`
if all selected onboarded signals have the same `dataSourceIds`. Ex:
- Signal A with dsId 12345, Signal B with dsId 12345, => store `[12345]`
- Signal A with dsId 12345, Signal B with dsId 67890, => do nothing instead of storing `[12345, 67890]`

And separately, I refactored the name `selectedDataSources` to
`selectedDataSourceIds`.

I will add unit tests once I go over this refactor with @elwang.

* AAM-41407: Skip unit test for now, will refactor and add more

* AAM-41407: [bug] Fix multi-create OB trait from "ALL" filter

When selecting multiple onboarded signals (with the same data source)
when the "ALL" signal type filter is selected, the create trait
button now correctly navigates to the Create Onboarded Trait page.

* AAM-41407: Add unit tests for new root-level selectors

* AAM-41407: Prevent "X Signals Selected" label from line-breaking